Cleaned the Core/SimpSMTSolvers a little.  Lookahead solver is now a subclass of SimpSMTSolver, but is untested.
