+++ src/api/MainSolver.C
@@ -387,18 +387,6 @@ sstat MainSolver::solve()
     if (!smt_solver->okay())
         return s_False;
 
-    if (config.sat_split_type() == spt_lookahead) {
-        int fix_vars;
-        if (config.sat_split_fixvars() > 0)
-            fix_vars = config.sat_split_fixvars();
-            fix_vars = getLog2Ceil(config.sat_split_num());
-        sstat res = lookaheadSplit(fix_vars);
-        return res;
-    }
-    else if (config.sat_pure_lookahead()) {
-        return lookaheadSplit(-1);
-    }
-
     vec<PTRef> query;
 
     vec<FrameId> en_frames;
