+++ src/smtsolvers/CoreSMTSolver.h
@@ -358,7 +358,7 @@ public:
     //
 protected:
     void  addVar    (Var v); // Ensure that var v exists in the solver
-    virtual Var newVar    (bool polarity = true, bool dvar = true); // Add a new variable with parameters specifying variable mode.
+    virtual Var newVar(bool polarity, bool dvar);//    (bool polarity = true, bool dvar = true); // Add a new variable with parameters specifying variable mode.
 public:
     bool    addOriginalClause(const vec<Lit> & ps);
     bool    addEmptyClause();                                   // Add the empty clause, making the solver contradictory.
@@ -374,11 +374,11 @@ public:
     bool    simplify     ();                        // Removes already satisfied clauses.
     void    declareVarsToTheories();                 // Declare the seen variables to the theories
     bool    solve        ( const vec< Lit > & assumps );                 // Search for a model that respects a given set of assumptions.
-    bool    solveLimited (const vec<Lit>& assumps); // Search for a model that respects a given set of assumptions (With resource constraints).
+//    bool    solveLimited (const vec<Lit>& assumps); // Search for a model that respects a given set of assumptions (With resource constraints).
     bool    solve        ();                        // Search without assumptions.
-    bool    solve        (Lit p);                   // Search for a model that respects a single assumption.
-    bool    solve        (Lit p, Lit q);            // Search for a model that respects two assumptions.
-    bool    solve        (Lit p, Lit q, Lit r);     // Search for a model that respects three assumptions.
+//    bool    solve        (Lit p);                   // Search for a model that respects a single assumption.
+//    bool    solve        (Lit p, Lit q);            // Search for a model that respects two assumptions.
+//    bool    solve        (Lit p, Lit q, Lit r);     // Search for a model that respects three assumptions.
     virtual bool  okay   () const;                  // FALSE means solver is in a conflicting state
     void    crashTest    (int, Var, Var);           // Stress test the theory solver
 
@@ -560,109 +560,6 @@ protected:
         bool operator()(Var v) const { return s.assigns[v] == l_Undef && s.decision[v]; }
     };
 
-    // Stuff specific to the lookahead implementation
-
-    uint32_t latest_round;                      // The numbering for arrays
-    void updateRound() { latest_round++; }
-    // -----------------------------------------------------------------------------------------
-    // Data type for exact value array
-    static inline int min(int i, int j) { return i < j ? i : j; }
-    static inline int max(int i, int j) { return i > j ? i : j; }
-
-    class ExVal
-    {
-    private:
-        int pprops;
-        int nprops;
-        int round;
-    public:
-        ExVal() : pprops(-1), nprops(-1), round(-1) {}
-        ExVal(int p, int n, int r) : pprops(p), nprops(n), round(r) {}
-        bool operator< (const ExVal& e) const
-        {
-            return (round < e.round) ||
-                   (min(pprops, nprops) < min(e.pprops, e.nprops)) ||
-                   ((min(pprops, nprops) == min(e.pprops, e.nprops)) && (max(pprops, nprops) < max(e.pprops, e.nprops)));
-        }
-        bool betterPolarity() const { return pprops < nprops; } // Should return false if the literal should be unsigned
-        int  getRound()       const { return round; }
-        int  getEx_l()        const { return min(pprops, nprops); }
-        int  getEx_h()        const { return max(pprops, nprops); }
-        void setRound(int r)        { round = r; }
-    };
-
-    // -----------------------------------------------------------------------------------------
-    // Data type for upper bound array
-    //
-    class UBel
-    {
-    public:
-        int ub;
-        int round;
-
-        UBel() : ub(-1), round(-1) {}
-        UBel(int u, int r) : ub(u), round(r) {}
-
-        bool operator== (const UBel& o) const { return ub == o.ub && round == o.round; }
-        bool operator!= (const UBel& o) const { return !(operator==(o)); }
-    };
-
-    static const CoreSMTSolver::UBel UBel_Undef;
-
-    class UBVal
-    {
-    private:
-        UBel ub_p;
-        UBel ub_n;
-        bool current(const ExVal& e) const { return ub_p.round == ub_n.round && ub_p.round == e.getRound(); }
-    public:
-        UBVal() {}
-        UBVal(const UBel& ub_pos, const UBel& ub_neg) : ub_p(ub_pos), ub_n(ub_neg) {}
-        void updateUB_p(const UBel& ub) { if (ub_p.round < ub.round || (ub_p.round == ub.round && ub_p.ub > ub.ub)) ub_p = ub; }
-        void updateUB_n(const UBel& ub) { if (ub_n.round < ub.round || (ub_n.round == ub.round && ub_n.ub > ub.ub)) ub_n = ub; }
-
-        bool safeToSkip(const ExVal& e) const;
-        const UBel& getUB_p()  const { return ub_p; }
-        const UBel& getUB_n()  const { return ub_n; }
-        bool betterThan(const ExVal& e) const;
-        const UBel& getLow() const
-        {
-            if (ub_p.round != ub_n.round) return UBel_Undef;
-            else return ub_p.ub < ub_n.ub ? ub_p : ub_n;
-        }
-        const UBel& getHigh() const
-        {
-            if (ub_p.round != ub_n.round) return UBel_Undef;
-            else return  ub_p.ub < ub_n.ub ? ub_n : ub_p;
-        }
-    };
-
-    // The result from the lookahead loop
-    enum class LALoopRes
-    {
-        sat,
-        unsat,
-        unknown,
-        splits,
-        restart
-    };
-
-    class laresult {
-    public:
-        enum result { tl_unsat, sat, restart, unsat, ok };
-    private:
-        result value;
-    public:
-        explicit laresult(result v) : value(v) {}
-        bool operator == (laresult o) const { return o.value == value; }
-        bool operator != (laresult o) const { return o.value != value; }
-    };
-
-    laresult la_tl_unsat;
-    laresult la_sat;
-    laresult la_restart;
-    laresult la_unsat;
-    laresult la_ok;
 
     // Solver state:
     //
@@ -681,14 +578,6 @@ protected:
     OccLists<Lit, vec<Watcher>, WatcherDeleted>  watches;          // 'watches[lit]' is a list of constraints watching 'lit' (will go there if literal becomes true).
     vec<lbool>          assigns;          // The current assignments (lbool:s stored as char:s).
     vec<bool>           var_seen;
-
-
-    bool betterThan_ub(const UBVal& ub, const ExVal& e) const;
-
-    void updateLABest(Var v);
-
-    vec<UBVal>          LAupperbounds;    // The current upper bounds
-    vec<ExVal>          LAexacts;         // The current exact values
     vec<char>           polarity;         // The preferred polarity of each variable.
     vec<char>           decision;         // Declares if a variable is eligible for selection in the decision heuristic.
 protected:
@@ -720,67 +609,6 @@ protected:
     vec<char>           prev_polarity;    // The previous polarity of each variable.
 #endif
 
-    int la_round;                         // Keeps track of the lookahead round (used in lower bounds)
-    class LABestLitBuf {
-        private:
-            int size;
-            vec<Pair<ExVal,Lit> > buf;
-            vec<lbool>& assigns;
-            inline lbool value(Lit p) const { return assigns[var(p)]^sign(p); }
-            bool randomize;
-            double rnd_seed;
-        public:
-            // Use 0 for random seed to disable randomization
-            LABestLitBuf(int sz, vec<lbool>& assigns, bool randomize, double rnd_seed)
-                : size(sz)
-                , assigns(assigns)
-                , randomize(randomize)
-                , rnd_seed(rnd_seed) {
-                for (int i = 0; i < size; i++)
-                    buf.push(Pair<ExVal,Lit>{ExVal(), lit_Undef});
-            }
-            void insert(Lit l, ExVal& val) {
-                int i;
-                for (i = 0; i < size; i++) {
-                    ExVal &buf_val = buf[i].first;
-                    Lit buf_l = buf[i].second;
-                    if ((buf_val < val) || (value(buf_l) != l_Undef))
-                        break;
-                }
-                if (i == size)
-                    return;
-
-                Pair<ExVal,Lit> new_next = buf[i];
-                buf[i++] = Pair<ExVal,Lit>{val,l};
-                for (; i < size; i++) {
-                    Pair<ExVal,Lit> tmp = buf[i];
-                    buf[i] = new_next;
-                    new_next = tmp;
-                }
-            }
-            Lit getLit(int i) {
-                assert(i < size);
-                assert(i >= 0);
-                for (int j = 0; j < size; j++) {
-                    if (i+j < size && buf[i+j].second != lit_Undef && value(buf[i+j].second) == l_Undef)
-                        return buf[i+j].second;
-                    if (i-j >= 0 && buf[i-j].second != lit_Undef && value(buf[i-j].second) == l_Undef)
-                        return buf[i-j].second;
-                }
-                return lit_Undef;
-            }
-            Lit getLit() {
-                if (randomize) {
-                    int i = CoreSMTSolver::irand(rnd_seed, size);
-                    return getLit(i);
-                }
-                else
-                    return getLit(0);
-            }
-            int getSize() { return size; }
-    };
-
-    LABestLitBuf buf_LABests;
 
     // Temporaries (to reduce allocation overhead). Each variable is prefixed by the method in which it is
     // used, exept 'seen' wich is used in several places.
@@ -824,7 +652,6 @@ protected:
     void     updateSplitState();                                                       // Update the state of the splitting machine.
     bool     scatterLevel();                                                           // Are we currently on a scatter level.
     bool     createSplit_scatter(bool last);                                           // Create a split formula and place it to the splits vector.
-    bool     createSplit_lookahead();                                                  // Does not change the formula
     bool     excludeAssumptions(vec<Lit>& neg_constrs);                                // Add a clause to the database and propagate
     void     insertVarOrder   (Var x);                                                 // Insert a variable in the decision order priority queue.
     Lit      pickBranchLit    ();                                                      // Return the next decision variable.
@@ -840,15 +667,10 @@ protected:
     bool     okContinue       ();                                                      // Check search termination conditions
     void     runPeriodics     ();                                                      // Run the periodic functions from searcha
     void     learntSizeAdjust ();                                                      // Adjust learnts size and print something
-    lbool    solve_           (int max_conflicts = 0);                                                      // Main solve method (assumptions given in 'assumptions').
+    virtual lbool    solve_   ();                                                      // Main solve method (assumptions given in 'assumptions').
     void     reduceDB         ();                                                      // Reduce the set of learnt clauses.
     void     removeSatisfied  (vec<CRef>& cs);                                         // Shrink 'cs' to contain only non-satisfied clauses.
     void     rebuildOrderHeap ();
-
-    void     updateLAUB       (Lit l, int props);                                      // Check the lookahead upper bound and update it if necessary
-    laresult lookahead_loop   (Lit& best, int &idx, ConflQuota &confl_quota);
-    void     setLAExact       (Var v, int pprops, int nprops);                         // Set the exact la value
-    lbool    LApropagate_wrapper(ConflQuota &confl_quota);
 
     // Maintaining Variable/Clause activity:
     //
@@ -1055,13 +877,6 @@ protected:
     set< PTRef >     interface_equalities;       // To check that we do not duplicate eij
     set< PTRef >     atoms_seen;                 // Some interface equalities may already exists in the formula
 
-    //
-    // Added by Grisha
-    // bit blasting heuristic data
-    //
-    vec<Var>*                  bb_siblings;
-    // FIXME Map<Var,vec<Var>*,VarHash> bb_vars; Grisha's stuff
-
     int                next_it_i;                  // Next index i
     int                next_it_j;                  // Next index j
     //
@@ -1138,8 +953,7 @@ protected:
     // Added Code
     //=================================================================================================
 public:
-    LALoopRes lookaheadSplit(int d, int &idx, ConflQuota confl_quota);
-    lbool     lookaheadSplit(int d);
+
     void    printTrace() const;
 
 protected:
@@ -1320,47 +1134,14 @@ inline bool     CoreSMTSolver::withinBudget() const
 // FIXME: after the introduction of asynchronous interrruptions the solve-versions that return a
 // pure bool do not give a safe interface. Either interrupts must be possible to turn off here, or
 // all calls to solve must return an 'lbool'. I'm not yet sure which I prefer.
-inline bool     CoreSMTSolver::solve()
-{
-    budgetOff();
-    assumptions.clear();
-    return solve_() == l_True;
-}
-inline bool     CoreSMTSolver::solve  (Lit p)
-{
-    budgetOff();
-    assumptions.clear();
-    assumptions.push(p);
-    return solve_() == l_True;
-}
-inline bool     CoreSMTSolver::solve  (Lit p, Lit q)
-{
-    budgetOff();
-    assumptions.clear();
-    assumptions.push(p);
-    assumptions.push(q);
-    return solve_() == l_True;
-}
-inline bool     CoreSMTSolver::solve  (Lit p, Lit q, Lit r)
-{
-    budgetOff();
-    assumptions.clear();
-    assumptions.push(p);
-    assumptions.push(q);
-    assumptions.push(r);
-    return solve_() == l_True;
-}
+
 inline bool     CoreSMTSolver::solve  (const vec<Lit>& assumps)
 {
     budgetOff();
     assumps.copyTo(assumptions);
     return solve_() == l_True;
 }
-inline bool    CoreSMTSolver::solveLimited(const vec<Lit>& assumps)
-{
-    assumps.copyTo(assumptions);
-    return solve_() == l_True;
-}
+
 inline bool     CoreSMTSolver::okay   () const { return ok; }
 
 inline void     CoreSMTSolver::toDimacs(const char* file)
